# [dynamo] Relax guard introduced when tracing `__call__` on user defined object